LayoutTests/media/modern-media-controls/tracks-button/tracks-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>TracksButton</code> class.");

const tracksButton = new TracksButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("tracksButton.element.localName", "button");
shouldBeTrue('tracksButton.element.classList.contains("icon")');
shouldBeTrue('tracksButton.element.classList.contains("tracks")');
shouldBeEqualToString("tracksButton.iconName", `${Icons.Tracks}`);

const fullscreenTracksButton = new TracksButton({ layoutTraits: LayoutTraits.macOS | LayoutTraits.Fullscreen });

scheduler.frameDidFire = function()
{
    if ([tracksButton, fullscreenTracksButton].some(button => button.width == 0))
        return;

    shouldBeTrue('tracksButton.element.style.webkitMaskImage.includes("macOS/media-selection@")');
    shouldBeTrue('fullscreenTracksButton.element.style.webkitMaskImage.includes("macOS/media-selection-fullscreen@")');
    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
